stock_brief/requirements.txt
streamlit==1.41.1
yfinance>=1.2.0
plotly==5.24.1
kaleido==0.2.1
requests==2.32.3
openai==1.59.7
fpdf2==2.8.2
python-dotenv==1.0.1
Pillow==11.1.0
python-docx==1.1.2
PyMuPDF>=1.27.0
